pyproject.toml
[project]
name = "adk-playground"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "google-adk>=1.24.1",
    "google-generativeai>=0.8.6",
    "litellm>=1.81.8",
    "psutil>=7.2.2",
    "python-dotenv>=1.2.1",
    "yfinance>=1.1.0",
]
